+++ src/definitions/monitoring/index.js
@@ -1,0 +1,9 @@
+export default {
+  url: 'GoogleStackdriverMonitoring-UptimeChecks(https://cloud.google.com/monitoring)',
+  verb: 'GET',
+  valid: () => true,
+  id: () => 0,
+  collection: 0,
+  reply: () => 'OK',
+  log: false,
+};
