+++ src/definitions/slack/has.js
@@ -1,0 +1,1 @@
+export default (obj, prop) => Object.prototype.hasOwnProperty.call(obj, prop);
